Feat: Optionally Specify ECS Cluster on Deploys
The LAMP project has a single ECS cluster with multiple objects deployed
to it. In order to use the deploy-ecs workflow, it needs to specify the
cluster rather than have the cluster name inferred from the application
name. If the cluster name is not specified, it falls back to the
application name to keep backwards compatibility.